Added green areas and changed road generation to differentiate road objects by name
Green areas are now beeing generated.
The road generator used to combine all roads that share a point to a single gameobject. Now it only combines roads with the same name. (Does not work for roads with missing names)
Added data cleanud.
Hide roads and rooftops in the project hierarchy to make it clearer.